音乐播放器功能已完善，在本地环境中无差错地运行，远程控制功能的代码已部分编写，未调试

+++ NetworkMusicPlayer/src/cn/cvte/music/MusicListAdapter.java
@@ -74,10 +74,10 @@ public class MusicListAdapter extends BaseAdapter{
 			viewHolder.size_tv.setText(MusicFile.UNKNOW);
 		}
 		if (!MusicFile.UNKNOW.equals(mi.duration)){
-			int time = Integer.parseInt(mi.duration);
+			long time = Long.parseLong(mi.duration);
 			time/=1000;
-			int minutes = time/60;
-			int seconds = time%60;
+			long minutes = time/60;
+			long seconds = time%60;
 			String timeStr = minutes+":"+String.format("%02d", seconds); 
 			viewHolder.duration_tv.setText(timeStr);
 		}else {
